refactor: 로직 정리 완료 	modified:   scripts/ui-sign.js 	modified:   scripts/utils-sign.js 	modified:   scripts/utils.js 	modified:   signin/index.js 	modified:   signup/index.js 	modified:   signup/ui.js 	deleted:    signup/utils.js
