+++ scripts/ui-sign.js
@@ -7,16 +7,16 @@
 //이메일, 비밀번호 input
 const emailInput = document.querySelector("#input-email");
 const passwordInput = document.querySelector("#input-password");
-const passwordCheckInput = document.querySelector("#input-password-check");
 
 //이메일, 비밀번호 error-message
 const emailErrorMessageElement = document.querySelector("#error-email");
 const passwordErrorMessageElement = document.querySelector("#error-password");
-const passwordCheckErrorMessageElement = document.querySelector("#error-password-check");
 
 //비밀번호 보이기 버튼
 const eyeBtn = document.querySelector("#password-eye");
-const eyeBtnCheck = document.querySelector("#password-check-eye");
+
+//form element
+const formElement = document.querySelector(".sign");
 
 /********************
  * UTILITY FUNCTION
@@ -50,18 +50,15 @@ function hidePassword() {
   imgElement.src = "../public/icon/eye-off.svg";
 }
 
-//eyeBtnCheck 비밀번호 확인 보이게
-function showPasswordCheck() {
-  const imgElement = document.querySelector("#password-check-eye .icon-eye");
-  passwordCheckInput.type = "text";
-  imgElement.src = "../public/icon/eye-on.svg";
-}
-
-//eyeBtnCheck 비밀번호 확인 안보이게
-function hidePasswordCheck() {
-  const imgElement = document.querySelector("#password-check-eye .icon-eye");
-  passwordCheckInput.type = "password";
-  imgElement.src = "../public/icon/eye-off.svg";
-}
-
-export {};
+export {
+  emailInput,
+  passwordInput,
+  emailErrorMessageElement,
+  passwordErrorMessageElement,
+  eyeBtn,
+  formElement,
+  showError,
+  hideError,
+  showPassword,
+  hidePassword,
+};
