+++ scripts/utils.js
@@ -1,5 +1,9 @@
 /* Common Utility Logic */
 
+/********************
+ * UTILITY FUNCTION
+ ********************/
+
 //인풋이 비어있는지 검사
 function isTextEmpty(inputValue) {
   if (inputValue === "") {
